requirements.txt
################################################################################
# Core Dependencies
################################################################################
# Using numpy 1.x due to compatibility issues with older libs (basicsr, numba)
numpy<2.0.0

# PyTorch Ecosystem (Ensure CUDA 12.1 compatibility manually if needed)
torch>=2.5.0
torchvision>=0.20.0

################################################################################
# Hugging Face & Diffusion Dependencies
################################################################################
transformers>=4.40.0
diffusers>=0.30.0
accelerate>=0.30.0
huggingface-hub>=0.23.0
peft>=0.11.0
compel>=2.0.0
safetensors>=0.4.0

################################################################################
# Computer Vision & Image Processing
################################################################################
opencv-python>=4.9.0
Pillow>=10.3.0
scipy>=1.13.0
scikit-image>=0.22.0

# Ultralytics YOLO (Latest stable)
ultralytics>=8.3.0

# MediaPipe & Protobuf (Restricted due to conflicts)
mediapipe==0.10.14
protobuf<5.0.0

# InsightFace & ONNX
insightface>=0.7.3
onnxruntime-gpu>=1.17.0; sys_platform != 'darwin'
onnxruntime>=1.17.0; sys_platform == 'darwin'

# Segmentation
segment-anything

################################################################################
# Restoration & Enhancement (GFPGAN, Basicsr)
################################################################################
basicsr>=1.4.2
gfpgan>=1.3.8
# CodeFormer dependency often tricky, gfpgan handles basicsr well usually

################################################################################
# ControlNet Preprocessors
################################################################################
controlnet_aux>=0.0.9

################################################################################
# UI & Utilities
################################################################################
PyQt6>=6.6.0
pyyaml>=6.0
requests>=2.31.0
tqdm>=4.66.0
pandas>=2.2.0
piexif>=1.1.3
nudenet>=3.3.0

# Windows Runtime (Recommended for ONNX/OpenCV stability on Windows)
msvc-runtime; sys_platform == 'win32'